+++ lib/fx/fx.h
@@ -1,0 +1,21 @@
+#ifndef FX_H
+#define FX_H
+
+#include <stdint.h>
+#include <avr/io.h>
+#include <util/delay.h>
+#include "../lcd/lcd.h"
+#include "../state/state.h"
+#include "../shift/shift.h"
+
+// run appropriate function based on button state
+void mode_switcher(state *st, shiftReg *sr, uint8_t lcd_on);
+
+ // all bits on/off
+void onoff(shiftReg *sr, state *st, int num_sr, int on, uint8_t lcd);
+
+// chaser effects
+void byte_chaser(shiftReg *sr, state *st, int num_sr, uint8_t rev, uint8_t lcd);
+void bit_chaser(shiftReg *sr, state *st, uint8_t num_sr, uint8_t rev);
+
+#endif